requirements.txt
PySide6>=6.6.0
paramiko>=3.4.0
psutil>=5.9.0
requests>=2.31.0
pysnmp-lextudio>=6.1.0
